fix: add support for stubby to handle stringified requests (#10)
We should always be sending the request `body` as a `JSON` string,
however, at the moment due to `stubby`, we work around stubby to provide
`body` as an `Object` this means that we are not stubbing the same
information that we send in production. This commit keeps the current
behaviour and adds new functionality to convert a `JSON` string to an
`Object` to test; this is to make sure we don't break existing tests that
are reliant on the old functionality.